+++ CHANGELOG.md
@@ -15,9 +15,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Concurrent OCR and segmentation
 - Concurrent page processing
 - CPU support - run with `docker compose up -f compose-cpu.yaml -d`
+- Python client - `pip install chunkr-ai`
 
 ### Changed
 - Combined all workers into a `task` worker. See [279](https://github.com/lumina-ai-inc/chunkr/issues/279)
+- Changed `expires_at` to `expires_in`
 
 ### Fixed
 - List items incorrect heuristics. See [276](https://github.com/lumina-ai-inc/chunkr/issues/276)
@@ -29,3 +31,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Deprecated `target_chunk_length`, you can now use `chunk_processing.target_length` instead 
 - Deprecated `structured_extraction.json_schema.type`
 - Deprecated `ocr_strategy.Off`
+- Deprecated `expires_at` in the Python client